Rename taskmanagement package to myfeature
