+++ walking-skeleton/src/main/java/com/example/application/myfeature/TaskRepository.java
@@ -1,4 +1,4 @@
-package com.example.application.taskmanagement;
+package com.example.application.myfeature;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
