+++ walking-skeleton/src/main/java/com/example/application/myfeature/TaskService.java
@@ -1,4 +1,4 @@
-package com.example.application.taskmanagement;
+package com.example.application.myfeature;
 
 //#if ui.framework == "hilla"
 import com.vaadin.flow.server.auth.AnonymousAllowed;
