+++ walking-skeleton/src/main/java/com/example/application/myfeature/package-info.java
@@ -1,0 +1,11 @@
+@NullMarked
+//#if ui.framework == "hilla"
+@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
+//#endif
+package com.example.application.myfeature;
+// TODO Remove this package once you have added real features
+
+import org.jspecify.annotations.NullMarked;
+//#if ui.framework == "hilla"
+import org.springframework.lang.NonNullApi;
+//#endif
