+++ walking-skeleton/src/main/java/com/example/application/myfeature/ui/TaskListView.java
@@ -1,8 +1,8 @@
-package com.example.application.taskmanagement.ui;
+package com.example.application.myfeature.ui;
 
 import com.example.application.base.ui.component.ViewToolbar;
-import com.example.application.taskmanagement.Task;
-import com.example.application.taskmanagement.TaskService;
+import com.example.application.myfeature.Task;
+import com.example.application.myfeature.TaskService;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
 import com.vaadin.flow.component.datepicker.DatePicker;
@@ -23,8 +23,8 @@ import java.util.Optional;
 
 import static com.vaadin.flow.spring.data.VaadinSpringDataHelpers.toSpringPageRequest;
 
-@Route("task-list-flow")
-@PageTitle("Task List Flow")
+@Route("")
+@PageTitle("Task List")
 @Menu(order = 0, icon = "vaadin:clipboard-check", title = "Task List Flow")
 class TaskListView extends Main {
