+++ walking-skeleton/src/test/java/com/example/application/myfeature/TaskServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.application.taskmanagement;
+package com.example.application.myfeature;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
